feat: 회원 프로필 조회 api 추가

+++ src/main/java/com/example/demo/common/security/JwtInfo.java
@@ -1,4 +1,4 @@
 package com.example.demo.common.security;
 
-public record JwtInfo(Integer customerId, JwtType jwtType) {
+public record JwtInfo(Long customerId, JwtType jwtType) {
 }
